+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.5.1] - 2020-08-xx
+## [0.6.0] - 2020-08-18
+
+### Added
+
+- `android_content.Intent#package` property
+- `android_content.Intent#startActivity()` method
 
 ## [0.5.0] - 2020-08-01
 
@@ -277,7 +282,7 @@ No functional changes.
 - `android_content.Context.packageName` getter
 - `android_content.Context.packageResourcePath` getter
 
-[0.5.1]:  https://github.com/drydart/flutter_android/compare/0.5.0...0.5.1
+[0.6.0]:  https://github.com/drydart/flutter_android/compare/0.5.0...0.6.0
 [0.5.0]:  https://github.com/drydart/flutter_android/compare/0.4.0...0.5.0
 [0.4.0]:  https://github.com/drydart/flutter_android/compare/0.3.7...0.4.0
 [0.3.7]:  https://github.com/drydart/flutter_android/compare/0.3.6...0.3.7